made separate weather and movies components and brought the components into the server.js file. Tested and it is functioning.
